pyproject.toml
[tool.poetry]
name = "little-conf"
version = "0.0.0"  # update automatically by CI
description = "Simple library for service configuration"
authors = ["smartleks"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = ">5.0.0"
pytest = "^6.2.4"
pydantic = "^1.8.0"
deepmerge = "^1.1.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py37"]
skip-string-normalization = true
